Create Find coding sequences
Enter prokaryotic sequences and I will scout them for possible ATG codons and stop codons. I will print in the output the position of the possible ATG codons and I will return the phrase: This is a coding sequence, if any by printing the position of the first stop codon found.
If I will not find any coding sequence I will return the length of the sequence of interest and the number of ATG found.